+++ app/Http/Controllers/StudentsController.php
@@ -134,7 +134,11 @@ class StudentsController extends Controller
         $student->email = $request->input('email');
         $student->phone = $request->input('phone');
         if ($request->hasFile('img')) {
-            Storage::delete('public/uploads/' . $student->img);
+            if ($student->img != 'default.png') {
+                //delete img
+                Storage::delete('public/uploads/' . $student->img);
+
+            }
             $student->img = $this->imageValidate($request);;
         }
         $student->save();
